catalyst_count/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bokeh==3.5.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
Django==5.0.7
django-allauth==0.63.6
django-bootstrap4==24.3
django-environ==0.11.2
djangorestframework==3.15.2
fsspec==2024.6.1
Jinja2==3.1.4
locket==1.0.0
lz4==4.3.3
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.1
packaging==24.1
pandas==2.2.2
partd==1.4.2
pillow==10.4.0
psutil==6.0.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
tblib==3.0.0
toolz==0.12.1
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
xyzservices==2024.6.0
zict==3.0.0